multiple function refactor Thurs 15:37 Dec 15

+++ srcs/ft_strlen.c
@@ -14,12 +14,12 @@
 
 size_t		ft_strlen(const char *str)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	if (!str[i])
 		return (0);
 	while (str[i])
 		i++;
-	return((size_t)i);
+	return(i);
 }
